Move the `ValuesForTest.h` file into the `src` directory
It is also used outside of unit tests.
Also use absolute includes for files from the `src` directory consistently when they are included by unit tests.
