+++ test/ForwardTest.cpp
@@ -4,7 +4,7 @@
 
 #include <gtest/gtest.h>
 
-#include "../src/util/Forward.h"
+#include "util/Forward.h"
 
 template <typename Expected, typename T>
 void tester(T&& t) {
