+++ test/SynchronizedTest.cpp
@@ -4,7 +4,7 @@
 
 #include <gtest/gtest.h>
 
-#include "../src/util/Synchronized.h"
+#include "util/Synchronized.h"
 
 using ad_utility::Synchronized;
