+++ test/TupleHelpersTest.cpp
@@ -7,7 +7,7 @@
 #include <array>
 #include <string>
 
-#include "../src/util/TupleHelpers.h"
+#include "util/TupleHelpers.h"
 
 using namespace ad_tuple_helpers;
 using namespace std::literals;
