+++ test/engine/BindTest.cpp
@@ -7,8 +7,8 @@
 #include "../util/IdTableHelpers.h"
 #include "../util/IndexTestHelpers.h"
 #include "../util/OperationTestHelpers.h"
-#include "./ValuesForTesting.h"
 #include "engine/Bind.h"
+#include "engine/ValuesForTesting.h"
 #include "engine/sparqlExpressions/LiteralExpression.h"
 
 using namespace sparqlExpression;
