+++ test/engine/QueryExecutionTreeTest.cpp
@@ -6,12 +6,12 @@
 
 #include "../util/IdTableHelpers.h"
 #include "../util/IndexTestHelpers.h"
-#include "./ValuesForTesting.h"
 #include "engine/IndexScan.h"
 #include "engine/QueryExecutionTree.h"
 #include "engine/Sort.h"
 #include "engine/StripColumns.h"
 #include "engine/Union.h"
+#include "engine/ValuesForTesting.h"
 
 using namespace ad_utility::testing;
