+++ test/engine/StripColumnsTest.cpp
@@ -14,9 +14,9 @@
 #include "../util/IdTableHelpers.h"
 #include "../util/IndexTestHelpers.h"
 #include "../util/OperationTestHelpers.h"
-#include "./ValuesForTesting.h"
 #include "engine/QueryExecutionTree.h"
 #include "engine/StripColumns.h"
+#include "engine/ValuesForTesting.h"
 
 namespace {
 using Vars = std::vector<std::optional<Variable>>;
